// practice/react-app/api/api-client-generated/api.ts
export interface Character {
    /**
     * The id of the character.
     */
    'id'?: number;
    /**
     * The name of the character.
     */
    'name'?: string;
    /**
     * The status of the character (Alive, Dead or unknown).
     */
    'status'?: string;
    /**
     * The species of the character.
     */
    'species'?: string;
    /**
     * The type or subspecies of the character.
     */
    'type'?: string;
    /**
     * The gender of the character.
     */
    'gender'?: string;
    'origin'?: CharacterOrigin;
    'location'?: CharacterLocation;
    /**
     * Link to the character\'s image.
     */
    'image'?: string;
    'episode'?: Array<string>;
    /**
     * Link to the character\'s own URL endpoint.
     */
    'url'?: string;
    /**
     * Time at which the character was created in the database.
     */
    'created'?: string;
}
export interface CharacterListResponse {
    'info'?: CharacterListResponseInfo;
    'results'?: Array<Character>;
}
export interface CharacterListResponseInfo {
    /**
     * The total number of characters.
     */
    'count'?: number;
    /**
     * The total number of pages.
     */
    'pages'?: number;
    /**
     * URL of the next page.
     */
    'next'?: string | null;
    /**
     * URL of the previous page.
     */
    'prev'?: string | null;
}
export interface CharacterLocation {
    /**
     * The name of the character\'s last known location.
     */
    'name'?: string;
    /**
     * Link to the character\'s last known location.
     */
    'url'?: string;
}
export interface CharacterOrigin {
    /**
     * The name of the character\'s origin location.
     */
    'name'?: string;
    /**
     * Link to the character\'s origin location.
     */
    'url'?: string;
}